+++ lib/home/page/home_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('HomePage'),
+      ),
+    );
+  }
+}
